requirements.txt
streamlit
google-generativeai
reportlab
markdown